update init script

+++ bin/init.sh
@@ -10,12 +10,17 @@ redis-server --daemonize yes
 #############
 
 # install traefik depdendency
-mkdir temp
-curl --silent -L -o traefik.tar.gz https://github.com/traefik/traefik/releases/download/v2.6.3/traefik_v2.6.3_linux_amd64.tar.gz
-tar -xf traefik.tar.gz -C temp 2>&1 1>/dev/null
-mv ./temp/traefik . 
-rm -rf temp
-rm traefik.tar.gz
+if [ ! -f "./traefik" ]
+then
+    mkdir temp
+    curl --silent -L -o traefik.tar.gz https://github.com/traefik/traefik/releases/download/v2.6.3/traefik_v2.6.3_linux_amd64.tar.gz
+    tar -xf traefik.tar.gz -C temp 2>&1 1>/dev/null
+    mv ./temp/traefik . 
+    rm -rf temp
+    rm traefik.tar.gz
+else
+    echo "'traefik' binary already exists"
+fi
 
 ####################
 # leaderboard app  #
@@ -23,7 +28,12 @@ rm traefik.tar.gz
 
 
 # build standalone app
-pyinstaller --onefile ./bin/python/getTop10.py
+if [ ! -f "./dist/getTop10" ]
+then
+    pyinstaller --onefile ./bin/python/getTop10.py
+else
+    echo -e "'getTop10' leaderboard app already exists.\n\tRun 'pyinstaller --onefile ./bin/python/getTop10.py' to rebuild."
+fi
 
 # load crontab
 crontab ./bin/cron.leaderboard.txt && mkdir -p ./var/log
